+++ src/client.ts
@@ -143,6 +143,7 @@ import {
   PollVote,
   PollVoteData,
   PollVotesAPIResponse,
+  PushPreference,
   PushProvider,
   PushProviderConfig,
   PushProviderID,
@@ -191,6 +192,7 @@ import {
   TestSNSDataInput,
   TestSQSDataInput,
   TokenOrProvider,
+  TranslateResponse,
   UnBanUserOptions,
   UpdateChannelOptions,
   UpdateChannelResponse,
@@ -202,6 +204,7 @@ import {
   UpdatePollAPIResponse,
   UpdatePollOptionAPIResponse,
   UpdateSegmentData,
+  UpsertPushPreferencesResponse,
   UserCustomEvent,
   UserFilters,
   UserOptions,
@@ -298,7 +301,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     // set the key
     this.key = key;
     this.listeners = {};
-    this.state = new ClientState<StreamChatGenerics>();
+    this.state = new ClientState<StreamChatGenerics>({ client: this });
     // a list of channels to hide ws events from
     this.mutedChannels = [];
     this.mutedUsers = [];
@@ -321,6 +324,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       withCredentials: false, // making sure cookies are not sent
       warmUp: false,
       recoverStateOnReconnect: true,
+      disableCache: false,
       ...inputOptions,
     };
 
@@ -825,7 +829,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     // ensure we no longer return inactive channels
     this.activeChannels = {};
     // reset client state
-    this.state = new ClientState();
+    this.state = new ClientState({ client: this });
     // reset thread manager
     this.threads.resetState();
     // reset token manager
@@ -1234,10 +1238,12 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     for (const channelID in refMap) {
       const channel = this.activeChannels[channelID];
-      const state = channel.state;
+      if (channel) {
+        const state = channel.state;
 
-      /** deleted the messages from this user. */
-      state?.deleteUserMessages(user, hardDelete);
+        /** deleted the messages from this user. */
+        state?.deleteUserMessages(user, hardDelete);
+      }
     }
   };
 
@@ -1675,6 +1681,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       c.data = channelState.channel;
       c.offlineMode = offlineMode;
       c.initialized = !offlineMode;
+      c.push_preferences = channelState.push_preferences;
 
       let updatedMessagesSet;
       if (skipInitialization === undefined) {
@@ -1815,6 +1822,17 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   /**
+   * setPushPreferences - Applies the list of push preferences.
+   *
+   * @param {PushPreference[]} A list of push preferences.
+   *
+   * @return {<UpsertPushPreferencesResponse>}
+   */
+  async setPushPreferences(preferences: PushPreference[]) {
+    return await this.post<UpsertPushPreferencesResponse>(this.baseURL + '/push_preferences', { preferences });
+  }
+
+  /**
    * removeDevice - Removes the device with the given id. Clientside users can only delete their own devices
    *
    * @param {string} id The device id
@@ -1853,7 +1871,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   _addChannelConfig({ cid, config }: ChannelResponse<StreamChatGenerics>) {
-    this.configs[cid] = config;
+    if (this._cacheEnabled()) {
+      this.configs[cid] = config;
+    }
   }
 
   /**
@@ -1965,7 +1985,10 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     // For the time being set the key as membersStr, since we don't know the cid yet.
     // In channel.query, we will replace it with 'cid'.
-    this.activeChannels[tempCid] = channel;
+    if (this._cacheEnabled()) {
+      this.activeChannels[tempCid] = channel;
+    }
+
     return channel;
   };
 
@@ -1992,7 +2015,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     // only allow 1 channel object per cid
     const cid = `${channelType}:${channelID}`;
-    if (cid in this.activeChannels && !this.activeChannels[cid].disconnected) {
+    if (cid in this.activeChannels && this.activeChannels[cid] && !this.activeChannels[cid].disconnected) {
       const channel = this.activeChannels[cid];
       if (Object.keys(custom).length > 0) {
         channel.data = { ...channel.data, ...custom };
@@ -2001,7 +2024,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       return channel;
     }
     const channel = new Channel<StreamChatGenerics>(this, channelType, channelID, custom);
-    this.activeChannels[channel.cid] = channel;
+    if (this._cacheEnabled()) {
+      this.activeChannels[channel.cid] = channel;
+    }
 
     return channel;
   };
@@ -2533,6 +2558,23 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   /**
+   * translate - translates the given text to provided language
+   *
+   * @param {string} text
+   * @param {string} destination_language
+   * @param {string} source_language
+   *
+   * @return {TranslateResponse} Response that includes the message
+   */
+  async translate(text: string, destination_language: string, source_language: string) {
+    return await this.post<APIResponse & TranslateResponse>(this.baseURL + `/translate`, {
+      text,
+      source_language,
+      destination_language,
+    });
+  }
+
+  /**
    * _normalizeExpiration - transforms expiration value into ISO string
    * @param {undefined|null|number|string|Date} timeoutOrExpirationDate expiration date or timeout. Use number type to set timeout in seconds, string or Date to set exact expiration date
    */
@@ -2770,8 +2812,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    *
    * @returns {{ threads: Thread<StreamChatGenerics>[], next: string }} Returns the list of threads and the next cursor.
    */
-  async queryThreads(options?: QueryThreadsOptions) {
-    const opts = {
+  async queryThreads(options: QueryThreadsOptions = {}) {
+    const optionsWithDefaults = {
       limit: 10,
       participant_limit: 10,
       reply_limit: 3,
@@ -2779,11 +2821,16 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       ...options,
     };
 
-    const res = await this.post<QueryThreadsAPIResponse<StreamChatGenerics>>(this.baseURL + `/threads`, opts);
+    const response = await this.post<QueryThreadsAPIResponse<StreamChatGenerics>>(
+      `${this.baseURL}/threads`,
+      optionsWithDefaults,
+    );
 
     return {
-      threads: res.threads.map((thread) => new Thread({ client: this, threadData: thread })),
-      next: res.next,
+      threads: response.threads.map(
+        (thread) => new Thread<StreamChatGenerics>({ client: this, threadData: thread }),
+      ),
+      next: response.next,
     };
   }
 
@@ -2800,22 +2847,22 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    */
   async getThread(messageId: string, options: GetThreadOptions = {}) {
     if (!messageId) {
-      throw Error('Please specify the message id when calling partialUpdateThread');
+      throw Error('Please specify the messageId when calling getThread');
     }
 
-    const opts = {
+    const optionsWithDefaults = {
       participant_limit: 100,
       reply_limit: 3,
       watch: true,
       ...options,
     };
 
-    const res = await this.get<GetThreadAPIResponse<StreamChatGenerics>>(
-      this.baseURL + `/threads/${encodeURIComponent(messageId)}`,
-      opts,
+    const response = await this.get<GetThreadAPIResponse<StreamChatGenerics>>(
+      `${this.baseURL}/threads/${encodeURIComponent(messageId)}`,
+      optionsWithDefaults,
     );
 
-    return new Thread<StreamChatGenerics>({ client: this, threadData: res.thread });
+    return new Thread<StreamChatGenerics>({ client: this, threadData: response.thread });
   }
 
   /**
@@ -2843,6 +2890,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       'reply_count',
       'participants',
       'channel',
+      'custom',
     ];
 
     for (const key in { ...partialThreadObject.set, ...partialThreadObject.unset }) {
@@ -2854,7 +2902,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     }
 
     return await this.patch<GetThreadAPIResponse<StreamChatGenerics>>(
-      this.baseURL + `/threads/${encodeURIComponent(messageId)}`,
+      `${this.baseURL}/threads/${encodeURIComponent(messageId)}`,
       partialThreadObject,
     );
   }
@@ -2873,6 +2921,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    * _isUsingServerAuth - Returns true if we're using server side auth
    */
   _isUsingServerAuth = () => !!this.secret;
+
+  _cacheEnabled = () => !this._isUsingServerAuth() || !this.options.disableCache;
 
   _enrichAxiosOptions(
     options: AxiosRequestConfig & { config?: AxiosRequestConfig } = {
@@ -3067,26 +3117,73 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     });
   }
 
+  /**
+   * Creates a new block list
+   *
+   * @param {BlockList} blockList - The block list to create
+   * @param {string} blockList.name - The name of the block list
+   * @param {string[]} blockList.words - List of words to block
+   * @param {string} [blockList.team] - Team ID the block list belongs to
+   *
+   * @returns {Promise<APIResponse>} The server response
+   */
   createBlockList(blockList: BlockList) {
     return this.post<APIResponse>(`${this.baseURL}/blocklists`, blockList);
   }
 
-  listBlockLists() {
-    return this.get<APIResponse & { blocklists: BlockListResponse[] }>(`${this.baseURL}/blocklists`);
+  /**
+   * Lists all block lists
+   *
+   * @param {Object} [data] - Query parameters
+   * @param {string} [data.team] - Team ID to filter block lists by
+   *
+   * @returns {Promise<APIResponse & {blocklists: BlockListResponse[]}>} Response containing array of block lists
+   */
+  listBlockLists(data?: { team?: string }) {
+    return this.get<APIResponse & { blocklists: BlockListResponse[] }>(`${this.baseURL}/blocklists`, data);
   }
 
-  getBlockList(name: string) {
+  /**
+   * Gets a specific block list
+   *
+   * @param {string} name - The name of the block list to retrieve
+   * @param {Object} [data] - Query parameters
+   * @param {string} [data.team] - Team ID that blocklist belongs to
+   *
+   * @returns {Promise<APIResponse & {blocklist: BlockListResponse}>} Response containing the block list
+   */
+  getBlockList(name: string, data?: { team?: string }) {
     return this.get<APIResponse & { blocklist: BlockListResponse }>(
       `${this.baseURL}/blocklists/${encodeURIComponent(name)}`,
+      data,
     );
   }
 
-  updateBlockList(name: string, data: { words: string[] }) {
+  /**
+   * Updates an existing block list
+   *
+   * @param {string} name - The name of the block list to update
+   * @param {Object} data - The update data
+   * @param {string[]} data.words - New list of words to block
+   * @param {string} [data.team] - Team ID that blocklist belongs to
+   *
+   * @returns {Promise<APIResponse>} The server response
+   */
+  updateBlockList(name: string, data: { words: string[]; team?: string }) {
     return this.put<APIResponse>(`${this.baseURL}/blocklists/${encodeURIComponent(name)}`, data);
   }
 
-  deleteBlockList(name: string) {
-    return this.delete<APIResponse>(`${this.baseURL}/blocklists/${encodeURIComponent(name)}`);
+  /**
+   * Deletes a block list
+   *
+   * @param {string} name - The name of the block list to delete
+   * @param {Object} [data] - Query parameters
+   * @param {string} [data.team] - Team ID that blocklist belongs to
+   *
+   * @returns {Promise<APIResponse>} The server response
+   */
+  deleteBlockList(name: string, data?: { team?: string }) {
+    return this.delete<APIResponse>(`${this.baseURL}/blocklists/${encodeURIComponent(name)}`, data);
   }
 
   exportChannels(request: Array<ExportChannelRequest>, options: ExportChannelOptions = {}) {