+++ test/unit/channel.js
@@ -1257,7 +1257,22 @@ describe('Channel _initializeState', () => {
 });
 
 describe('Channel.query', async () => {
-	it('should not update pagination for queried message set', async () => {
+	it('should not populate client.activeChannels if caching is disabled', async () => {
+		const client = await getClientWithUser();
+		client._cacheEnabled = () => false;
+		const channel = client.channel('messaging', uuidv4());
+		const mockedChannelQueryResponse = {
+			...mockChannelQueryResponse,
+			messages: Array.from({ length: DEFAULT_QUERY_CHANNEL_MESSAGE_LIST_PAGE_SIZE }, generateMsg),
+		};
+		const mock = sinon.mock(client);
+		mock.expects('post').returns(Promise.resolve(mockedChannelQueryResponse));
+		await channel.query();
+		expect(Object.keys(client.activeChannels).length).to.be.equal(0);
+		mock.restore();
+	});
+
+	it('should update pagination for queried message set to prevent more pagination', async () => {
 		const client = await getClientWithUser();
 		const channel = client.channel('messaging', uuidv4());
 		const mockedChannelQueryResponse = {
@@ -1268,11 +1283,11 @@ describe('Channel.query', async () => {
 		mock.expects('post').returns(Promise.resolve(mockedChannelQueryResponse));
 		await channel.query();
 		expect(channel.state.messageSets.length).to.be.equal(1);
-		expect(channel.state.messageSets[0].pagination).to.eql({ hasNext: true, hasPrev: true });
+		expect(channel.state.messageSets[0].pagination).to.eql({ hasNext: false, hasPrev: true });
 		mock.restore();
 	});
 
-	it('should update pagination for queried message set to prevent more pagination', async () => {
+	it('should not update pagination for queried message set', async () => {
 		const client = await getClientWithUser();
 		const channel = client.channel('messaging', uuidv4());
 		const mockedChannelQueryResponse = {
@@ -1283,7 +1298,7 @@ describe('Channel.query', async () => {
 		mock.expects('post').returns(Promise.resolve(mockedChannelQueryResponse));
 		await channel.query();
 		expect(channel.state.messageSets.length).to.be.equal(1);
-		expect(channel.state.messageSets[0].pagination).to.eql({ hasNext: true, hasPrev: false });
+		expect(channel.state.messageSets[0].pagination).to.eql({ hasNext: false, hasPrev: false });
 		mock.restore();
 	});
 });